requirements.txt
streamlit==1.27.2
pandas==2.1.1
python-dotenv==1.0.0